% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{copy_table.default}
\alias{copy_table.default}
\title{Copy data out of R}
\usage{
\method{copy_table}{default}(obj, row.names = FALSE, size = 4096)
}
\arguments{
\item{obj}{data of type data.frame, zoo, xts or list}

\item{row.names}{should row names be included defaults to FALSE}

\item{size}{set directory size - leave default as 4KB}
}
\description{
Copy data out of R
}
