% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{paste_table}
\alias{paste_table}
\title{Paste and object from the clipboard}
\usage{
paste_table()
}
\value{
A \pkg{data.frame} of the data in the clipboard.
}
\description{
Returns the clipboard as a \pkg{data.frame}.
}
\author{
Nicholas Marshall
}
